simplify even more path to make travis happy

+++ tests/scripts/src/test/scripts/structure_test/structure_test.sh
@@ -31,7 +31,11 @@ then
   mkdir -p $destDir
 fi
 
-wget -O $fileName --no-verbose $scriptPath
+if [[ ! -f $fileName ]]
+then
+    wget -O $fileName --no-verbose $scriptPath
+fi
+
 chmod +x $fileName
 
 IMAGE=$1
@@ -42,7 +46,7 @@ TEST_IMAGE="${IMAGE}-struct-test"
 pushd $5
 echo "Creating temporary image $TEST_IMAGE"
 pwd
-docker build --build-arg=RUNTIME_IMAGE=$IMAGE -f Dockerfile -t $TEST_IMAGE $WORKSPACE
+docker build --build-arg=RUNTIME_IMAGE=$IMAGE -f Dockerfile -t $TEST_IMAGE workspace
 popd
 
 $fileName test --image $TEST_IMAGE --config $CONFIG
